% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pf3k_SNP_AD_extract.R
\name{AD_mat_creation}
\alias{AD_mat_creation}
\title{Allele Depth Matrix Creation}
\usage{
AD_mat_creation(AD, lookup)
}
\arguments{
\item{AD}{Output of \code{Pf3k_SNP_AD_extraction}}

\item{lookup}{object. List of lists, with each list having 2 elements
chromosome and positions.}
}
\value{
returns 1 list of 2 matrices representing the REF and MAF frequencies
required
}
\description{
This function returns list of 2 matrices corrresponding to the allele depth for
REF and MAF, calculated from the output of \code{Pf3k_SNP_AD_extraction}
}
