% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignGTforREALMcCOIL}
\alias{assignGTforREALMcCOIL}
\title{Assign genotype calls based on a wsaf matrix}
\usage{
assignGTforREALMcCOIL(wsaf, err = 0.1)
}
\arguments{
\item{wsaf}{Within sample alle frequency matrix    An R data frame of SNP calling information. Row names are names of samples and column names are names of assays.}

\item{err}{Error in sequencing for determining if is actually a 
heterozygous call. Default = 0.1}
}
\description{
This function triggers the c code for the categorical method
}
