% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{McCOIL_categorical_cpp}
\alias{McCOIL_categorical_cpp}
\title{McCOIL_categorical_cpp}
\usage{
McCOIL_categorical_cpp(paramList)
}
\arguments{
\item{paramList}{A list of parameters created with equivalent R function}
}
\description{
Categorical cpp code
}
\details{
\code{McCOIL_categorical_cpp} implements THE REAL McCOIL categorical method
}
