% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCOIL_proportional_2018.R
\name{McCOIL_proportional_2018_R}
\alias{McCOIL_proportional_2018_R}
\title{The Real McCOIL proportional method function}
\usage{
McCOIL_proportional_2018_R(dataA1, dataA2, maxCOI = 25,
  totalrun = 10000, burnin = 1000, M0 = 15, epsilon = 0.02,
  err_method = 1, path = getwd(), saves = 10)
}
\arguments{
\item{dataA1}{The intensity of signals of allele 1 from the SNP assay. Row names are names of samples and column names are names of assays.}

\item{dataA2}{The intensity of signals of allele 2 from the SNP assay. Row names are names of samples and column names are names of assays.}

\item{maxCOI}{Upper bound for COI. The default is 25.}

\item{totalrun}{The total number of MCMC iterations. The default is 10000.}

\item{burnin}{The total number of burnin iterations. The default is 1000.}

\item{M0}{Initial COI. The default is 15.}

\item{epsilon}{The level of measurement error (eest). The default is 0.2.}

\item{err_method}{The default is 1. 
1: use pre-specified epsilon; 
2: use likelihood-free sampling for epsilon; 
3: update epsilon according to likelihood (for 2 and 3, pre-specified epsilon was used as initial value)}

\item{path}{The default is the current directory.}

\item{saves}{After how many iterations do we save/write outputs. Default = 10}

\item{output}{The name of output file. The default is output.txt.}
}
\value{
summary of output as data.frame
}
\description{
This function triggers the c code for the proportional method, where we 
now store thinned chains and return these rather than writing to a file.
}
