% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{McCOIL_proportional_cpp}
\alias{McCOIL_proportional_cpp}
\title{McCOIL_proportional_cpp}
\usage{
McCOIL_proportional_cpp(paramList)
}
\arguments{
\item{paramList}{A list of parameters created with equivalent R function}
}
\description{
Proportional cpp code
}
\details{
\code{McCOIL_proportional_cpp} implements THE REAL McCOIL proportional method
}

