% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pf3k_SNP_AD_extract.R
\name{Pf3k_SNP_AD_extraction}
\alias{Pf3k_SNP_AD_extraction}
\title{Pf3k personal REF and MAF AD extraction}
\usage{
Pf3k_SNP_AD_extraction(chromosome, positions,
  ftp_address = "ftp://ngs.sanger.ac.uk/production/pf3k/release_5/5.1/",
  qual = FALSE)
}
\arguments{
\item{chromosome}{Numeric giving chromosome number to search within}

\item{positions}{Numeric giving chromosome positions}

\item{ftp_address}{String for ftp url address. Default is "ftp://ngs.sanger.ac.uk/production/pf3k/release_5/5.1/"}

\item{qual}{Boolean detailing if QUAL for each SNP is extracted too. Default = FALSE}
}
\value{
returns 1 list of n lists, where n is length of positions, and each 
of the n lists contains 3 lists (REF & MAF & QUAL) of length m, where m is the number
of sequences within Pf3k. Where more than one type of MAF is possible, 
the most common MAF is recorded.
}
\description{
This function returns the allele depth for a given chromosme and position
from the Pf3k database
}
